Specs: Paginate `stub_get_project_assignments`
To force our Harvest::Api coverage to 100%.
We might add separate tests for it later, but this is an acceptable way to reach 100% in my opinion.
If we do not cover all provider a piece of code when testing all branches of every script, then the code should be removable.